+++ src/handlers/subscribe.rs
@@ -1,5 +1,5 @@
 use crate::{error::HandlerResult, messages::Message, LongPollingServiceContext};
-use axum::{extract::State, Json};
+use axum::{extract::State, http::StatusCode, Json};
 use std::sync::Arc;
 
 pub(crate) async fn subscribe(
@@ -28,7 +28,11 @@ pub(crate) async fn subscribe(
         let client_id =
             client_id.ok_or_else(|| Message::session_unknown(id.clone(), channel.clone(), None))?;
 
-        let validate = |name: &String| context.channel_name_validator().validate(name);
+        let validate = |name: &String| {
+            context
+                .channel_name_validator()
+                .validate_error(name, StatusCode::BAD_REQUEST)
+        };
         subscription.iter().try_for_each(validate)?;
 
         context