+++ src/utils.rs
@@ -1,3 +1,5 @@
 mod channel_name_validation;
+#[cfg(test)]
+mod test;
 
 pub(crate) use channel_name_validation::*;